atf-juno: deploy juno r2 dtb on older kernels
We ensure juno r2 dtb is always present in the recovery image.
It's also needed when we boot from TFTP.
Ensure to deploy Juno r2 dtb in any case. When it doesn't exist (on 4.4
kernel for example), copy Juno r1 dtb.

+++ recipes-bsp/atf/atf-juno_git.bb
@@ -63,9 +63,16 @@ do_deploy() {
     cp -aL ${DEPLOY_DIR_IMAGE}/Image-juno-r1.dtb \
     ${WORKDIR}/juno-oe-uboot/SOFTWARE/juno-r2.dtb
 
-    [ -L ${DEPLOY_DIR_IMAGE}/Image-juno-r2.dtb ] && \
-    cp -aL ${DEPLOY_DIR_IMAGE}/Image-juno-r2.dtb \
-    ${WORKDIR}/juno-oe-uboot/SOFTWARE/juno-r2.dtb
+    # Overwrite Juno dtb with the real r2 dtb in the recovery image
+    if [ -L ${DEPLOY_DIR_IMAGE}/Image-juno-r2.dtb ]; then
+      cp -aL ${DEPLOY_DIR_IMAGE}/Image-juno-r2.dtb \
+      ${WORKDIR}/juno-oe-uboot/SOFTWARE/juno-r2.dtb
+    else
+      # Deploy a dummy r2 dtb, copy of Juno r1 dtb
+      dt_src=$(readlink ${DEPLOY_DIR_IMAGE}/Image-juno-r1.dtb)
+      dt_dst=${dt_src/juno-r1/juno-r2}
+      cp -a ${DEPLOY_DIR_IMAGE}/${dt_src} ${DEPLOY_DIR_IMAGE}/${dt_dst}
+    fi
 
     # Move the ramdisk up in NOR flash to give more space for a larger kernel.
     # This also means that we have less space for the ramdisk, however, on OE
